pause menu on forrest
