+++ Assets/Animations/Menu.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5cedf8830ffa69645a5fe10bab9bc091
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
