+++ Assets/Animations/Menu/Panel.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9812f2521a1b38b419281ac399ea5291
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
