+++ Assets/Animations/Menu/PauseMenu.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5a693ed4146e4424daec463e4f79080e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
