+++ Assets/Scenes/Menu.unity
@@ -189,7 +189,7 @@ RectTransform:
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0, y: 0}
+  m_AnchorMax: {x: 0, y: 1}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 10, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
@@ -266,7 +266,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -417.3, y: -236.8}
+  m_AnchoredPosition: {x: -450, y: -260}
   m_SizeDelta: {x: 100, y: 40}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &178583763
@@ -553,7 +553,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!224 &413147278
 RectTransform:
   m_ObjectHideFlags: 0
@@ -610,7 +610,7 @@ RectTransform:
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0, y: 0}
+  m_AnchorMax: {x: 0, y: 1}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 20, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
@@ -685,8 +685,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 1, y: 1}
-  m_AnchoredPosition: {x: -414.6, y: -109.8}
-  m_SizeDelta: {x: -721.7999, y: -510}
+  m_AnchoredPosition: {x: -323.2, y: -73.3}
+  m_SizeDelta: {x: -881.54, y: -531.41}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &531741737
 MonoBehaviour:
@@ -775,7 +775,7 @@ MonoBehaviour:
   m_firstVisibleCharacter: 0
   m_useMaxVisibleDescender: 1
   m_pageToDisplay: 1
-  m_margin: {x: 117.99962, y: 0, z: 56.67755, w: 22.846634}
+  m_margin: {x: -1.0664062, y: 0, z: -4.8399353, w: 1.0178528}
   m_textInfo:
     textComponent: {fileID: 531741737}
     characterCount: 6
@@ -938,7 +938,7 @@ MonoBehaviour:
   m_fontSizeMin: 18
   m_fontSizeMax: 72
   m_fontStyle: 0
-  m_textAlignment: 258
+  m_textAlignment: 1026
   m_characterSpacing: 7
   m_wordSpacing: 0
   m_lineSpacing: 0
@@ -1096,7 +1096,7 @@ MonoBehaviour:
   m_fontSizeMin: 18
   m_fontSizeMax: 72
   m_fontStyle: 0
-  m_textAlignment: 258
+  m_textAlignment: 514
   m_characterSpacing: 7
   m_wordSpacing: 0
   m_lineSpacing: 0
@@ -1197,8 +1197,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -329.7, y: -132.6}
-  m_SizeDelta: {x: 250, y: 25}
+  m_AnchoredPosition: {x: -323.2, y: -110.2}
+  m_SizeDelta: {x: 292.68, y: 20.03}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &847673416
 MonoBehaviour:
@@ -1288,8 +1288,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -406.1, y: -12.2}
-  m_SizeDelta: {x: 130, y: 60}
+  m_AnchoredPosition: {x: -405.63, y: 0.000005722}
+  m_SizeDelta: {x: 127.13, y: 58.7}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1077503957
 MonoBehaviour:
@@ -1418,8 +1418,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 1, y: 1}
-  m_AnchoredPosition: {x: -299.59998, y: -10.200012}
-  m_SizeDelta: {x: -721.7999, y: -510}
+  m_AnchoredPosition: {x: -323.2, y: 5}
+  m_SizeDelta: {x: -707.32, y: -510}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1091446554
 MonoBehaviour:
@@ -1487,7 +1487,7 @@ MonoBehaviour:
   m_enableWordWrapping: 1
   m_wordWrappingRatios: 0.4
   m_overflowMode: 0
-  m_firstOverflowCharacterIndex: -1
+  m_firstOverflowCharacterIndex: 0
   m_linkedTextComponent: {fileID: 0}
   m_isLinkedTextComponent: 0
   m_isTextTruncated: 0
@@ -1508,7 +1508,7 @@ MonoBehaviour:
   m_firstVisibleCharacter: 0
   m_useMaxVisibleDescender: 1
   m_pageToDisplay: 1
-  m_margin: {x: -1.1475525, y: 0, z: 56.67755, w: -10.173355}
+  m_margin: {x: -1.1475525, y: 0, z: -0.46640015, w: -0.26835632}
   m_textInfo:
     textComponent: {fileID: 1091446554}
     characterCount: 7
@@ -1578,8 +1578,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -359.2, y: -85.5}
-  m_SizeDelta: {x: 220, y: 60}
+  m_AnchoredPosition: {x: -360.34, y: -78.6}
+  m_SizeDelta: {x: 217.71, y: 58.48}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1161713017
 MonoBehaviour:
@@ -1879,8 +1879,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -404.4, y: -165.5}
-  m_SizeDelta: {x: 130, y: 60}
+  m_AnchoredPosition: {x: -406.49, y: -155.6}
+  m_SizeDelta: {x: 125.42, y: 52.38}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1253564041
 MonoBehaviour:
@@ -1991,7 +1991,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!224 &1296289232
 RectTransform:
   m_ObjectHideFlags: 0
@@ -2193,7 +2193,7 @@ MonoBehaviour:
   m_fontSizeMin: 18
   m_fontSizeMax: 72
   m_fontStyle: 0
-  m_textAlignment: 258
+  m_textAlignment: 514
   m_characterSpacing: 7
   m_wordSpacing: 0
   m_lineSpacing: 0
