+++ Assets/Scenes/Menu/Violet Wasteland.otf.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 10dd9a39d718f684cb8b06d829ce89e7
+TrueTypeFontImporter:
+  externalObjects: {}
+  serializedVersion: 4
+  fontSize: 16
+  forceTextureCase: -2
+  characterSpacing: 0
+  characterPadding: 1
+  includeFontData: 1
+  fontName: Violet Wasteland
+  fontNames:
+  - Violet Wasteland
+  fallbackFontReferences: []
+  customCharacters: 
+  fontRenderingMode: 0
+  ascentCalculationMode: 1
+  useLegacyBoundsCalculation: 0
+  shouldRoundAdvanceValue: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
