+++ ProjectSettings/EditorBuildSettings.asset
@@ -12,9 +12,15 @@ EditorBuildSettings:
     path: Assets/Scenes/Test.unity
     guid: 8b18da02b00ef7f48b1c4cd497671301
   - enabled: 1
+    path: Assets/Scenes/Beggining_Scene.unity
+    guid: d186542acd38eb64c950182b96f7b86b
+  - enabled: 1
     path: Assets/Scenes/Village_Scene.unity
     guid: 720fbb8b113184c4f83533a5a7cf7408
   - enabled: 1
     path: Assets/Scenes/Forest_Scene.unity
     guid: 520c447c3cbbce14692ba32f621b7971
+  - enabled: 0
+    path: Assets/Scenes/level01.unity
+    guid: 72cf249264aad394a9bd22a3cc68e8d0
   m_configObjects: {}
